Add GitHub pages workflow for publishing JSON collections
This will publish the JSON collections to GitHub pages when new tags are
made or when the workflow is run manually. This will allow controlled
releases of the collections in a CDN fronted environment that's
appropriate for consumption in our apps at runtime.

The workflow is slightly customized from the static HTML starter
workflow suggested by GitHub. It uses an Environment rather than
publishing from a git branch. By default the environment will allow
deploying from `main`, but it also needs to be allowed to deploy from
the `v*` pattern for tags.